requirements.txt
feedparser
requests
beautifulsoup4
playwright
google-generativeai
python-dotenv
pytest
